[Compiler] Adapt SymbolTable to changed module API
This commit makes the SymbolTable parse module files and traverse the
AST instead of running the javascript. That is dumber, but the only
doable solution with the module API we decided on (creating properties
inside the constructor), as else we'd need to create an object in
order to learn about its properties, which would break too easily.
This stuff is now done in an own module class.