[Runtime] Change module API to simplify parsing.
From now on, modules must call __engine.registerModule(), passing the
module data to it, instead of returning it. It doesn't make a huge
difference for the programmer, developing a module, neither for the
runtime, but it makes it a lot easier for the compiler parsing the
module file to know which expression is the module data. Prior we
needed to look at all return statements and guess if it's the correct
one, now we can be sure.