[Compiler] ModuleLoader: Detect props and stuff.
Now ModuleLoader is able to detect properties, methods, signals and
inheritancies. That means it's able to detect the property and its
name and adds it to the class object (the same for methods and
signals). It's not yet able to detect attributes of properties like
defaultValue, readonly, etc. Also it's not able to detect parameters
of methods and signals.

All tests work.